реализована выдача статистических данных по эндпоинту GET /stats

+++ stat-service/src/main/java/explore/with/me/model/ViewStats.java
@@ -9,5 +9,5 @@ public class ViewStats {
 
     String app;
     String uri;
-    Long hits;
+    int hits;
 }
